change text length to uint32

+++ pc/cpp/test/test_decoder.cpp
@@ -6,7 +6,7 @@ TEST(Decoder_get_text, basic) {
     Decoder d;
 
     d.put_text("123456", 6);
-    uint16_t res = d.get_text(s);
+    uint32_t res = d.get_text(s);
 
     ASSERT_EQ(res, 6);
     ASSERT_STREQ(s, "123456");
@@ -21,7 +21,7 @@ TEST(Decoder_get_text, edge_length) {
     strncpy(predict_s, s, 65535);
 
     d.put_text(s, 65535);
-    uint16_t res = d.get_text(s);
+    uint32_t res = d.get_text(s);
 
     ASSERT_EQ(res, 65535);
     ASSERT_STREQ(s, predict_s);
